k8s: Move admin API mock to shared package
In the statefulset unit test the admin API needs to be mocked as cluster
health should be available.
